Add rules for box, inner and active classes
box adds aesthetic style rules to nested divs with class box
inner sets display: flex; to divs with class inner in order to display
divs with class box horizontally and enable a grid design.
active is a class that will be added the divs with class box using the
js script to enable the 'etch-a-sketch' look.