Add script to dynamically populate dom with a grid
The structure of the grid is:
mainContainer class is parent element <div>
innerContainer is nested child of mainContainer, <div>
box is nested child of innerContainer, <div>

The grid is built using a nested for loop. The inner loop creates box
divs and appends them to an innerContainer div on each iteration. After
inner loop finishes, the innerContainer is appended to the main
container and the outter loop iterates to allow the creation of another
innerContainer div with box divs. The size of the grid is established
by the constant BOXES.

Future goal is to set up event listener on each box, and add active
class on click in order to change aesthetic display and allow for
drawing on the grid.