Create uniform grid of boxes and change styles on event
Functions:
getRandomRGB() - returns random values for red, green and blue in order
to style a box in the grid when an event happens
Drawing by click:
    - addClickDrawing()
    - clickEventHandler()
    - removeClickDrawing()
Drawing by hover (i.e. mouseenter):
    - addmouseenterDrawing()
    - mouseenterEventHandler()
    - removemouseenterDrawing()
Reset default styles with resetBoxStyle()

Functions for click drawing are called onClick of click button,
for mouseenter drawing ... of hover button, and all drawing is reset
and stopped onClick of clear button.

Future goal is to let user select height and width (i.e. size of grid)
and to adjust grid size with JS accordingly

adjustedHeight of each box reasoning:
The grid container is always 500px by 500 px. The boxes, with flex-grow
set to 1, grow in the horizontal main-axis direction and adjust width
accordingly. By calculating adjustedHeight -- that is, getting the value
of width for a given box after flex-growing -- the height can be set to
match the width for uniformity all while remaining within and not
overflowing the grid container.